Refactor builder to use docker packages
Mainly `github.com/docker/docker/builder/dockerfile/parser` and probably
a few helper too.

It should be possible to use builder/dockerfile as is but needs some
stuff to adapt, another road to experiment at some point.
